% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bal.R
\name{bal}
\alias{bal}
\title{Function for fitting a Bayesian linear quantile regression with asymmetric Laplace (AL) errors 
       via Markov chain Monte Carlo (MCMC).}
\usage{
bal(
  resp,
  covars,
  prob,
  beta_prior,
  priors,
  starting,
  mcmc_settings,
  verbose = TRUE
)
}
\arguments{
\item{resp}{a vector of response data of length \eqn{n}}

\item{covars}{an \eqn{n\times p} matrix of \eqn{p} covariates.}

\item{prob}{a quantile probability with value in (0,1). For example, if prob is 0.5, the model corresponds to a median quantile regression.}

\item{beta_prior}{a quoted keyword that specifies the prior for regression parameter \eqn{\beta}.
Supported key words are \code{gaussian} and \code{laplace}.}

\item{priors}{a list of priors. Each element of the list corresponds to a combination of an unknown parameter
and its prior distribution; for example, "\code{phi_invgamma}".}

\item{starting}{a list of starting values. Each element of the list corresponds to an unknown parameter.}

\item{mcmc_settings}{a list of MCMC simulation parameters.
\code{n_iter}: number of iterations; \code{n_burn}: number of burn-in samples;
\code{n_thin}: thinning degree. If   \code{verbose = TRUE},
\code{n_report} defines the interval to report MCMC progress.}

\item{verbose}{logical; if true, model specification and MCMC progress are printed to the screen.}
}
\value{
An object of class \code{bal}. The return object is a list that comprises:

\tabular{ll}{

\code{post_sams} \tab a list of posterior samples. Each element of the list corresponds to an unknown parameter. \cr
\tab \cr
\code{response} \tab the response vector. \cr
\tab \cr
\code{covariates} \tab the covariate matrix. \cr
\tab \cr
\code{prob} \tab the quantile probability. \cr
\tab \cr
\code{beta_prior} \tab the prior for regression parameter \eqn{\beta}. \cr
\tab \cr
\code{starting} \tab a list of starting values. \cr
\tab \cr
\code{mcmc_settings}: \tab a list of MCMC simulation parameters:
       \code{n_iter}: number of iterations; 
       \code{n_burn}: number of burn-in samples;
       \code{n_thin}: thinning degree. If   \code{verbose = TRUE},
       \code{n_report} defines the interval to report MCMC progress and Metropolis acceptance rate (if \code{ga_sampler = metropolis}).
 \cr
\tab \cr
\code{runtime}: \tab running time for MCMC simulation calculated using \code{system.time}. \cr

}
}
\description{
Function for fitting a Bayesian linear quantile regression with asymmetric Laplace (AL) errors 
       via Markov chain Monte Carlo (MCMC).
}
\references{
Kozumi, H. and Kobayashi, G. (2011), 
"Gibbs sampling methods for Bayesian quantile regression." 
Journal of Statistical Computation and Simulation, 81, 1565–1578.
DOI: \href{https://doi.org/10.1080/00949655.2010.496117}{https://doi.org/10.1080/00949655.2010.496117}.
}
\author{
Xiaotian Zheng \email{xiaotianzheng42@gmail.com}
}
