% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgal.R
\name{bgal}
\alias{bgal}
\title{Function for fitting a Bayesian linear quantile regression with generalized asymmetric Laplace (GAL) errors
       via Markov chain Monte Carlo (MCMC).}
\usage{
bgal(
  resp,
  covars,
  prob,
  beta_prior,
  priors,
  starting,
  tuning,
  mcmc_settings,
  ga_sampler = "slice",
  verbose = TRUE
)
}
\arguments{
\item{resp}{a vector of response data of length \eqn{n}.}

\item{covars}{an \eqn{n\times p} matrix of \eqn{p} covariates.}

\item{prob}{a quantile probability with value in (0,1). For example, if prob is 0.5, the model corresponds to a median quantile regression.}

\item{beta_prior}{a quoted keyword that specifies the prior for regression parameter \eqn{\beta}.
Supported key words are \code{gaussian} and \code{laplace}.}

\item{priors}{a list of priors. Each element of the list corresponds to a combination of an unknown parameter
and its prior distribution; for example, "\code{phi_invgamma}".}

\item{starting}{a list of starting values. Each element of the list corresponds to an unknown parameter.}

\item{tuning}{a list of tuning parameters. Each element of the list corresponds to an unknown parameter.
For example, \code{step_size}: the step size or initial interval width 
of the stepping out procedure of the slice sampler; \code{se_ga}: the standard deviation of the random
walk Metropolis sampler proposal distribution.}

\item{mcmc_settings}{a list of MCMC simulation parameters.
\code{n_iter}: number of iterations; \code{n_burn}: number of burn-in samples;
\code{n_thin}: thinning degree. If   \code{verbose = TRUE},
\code{n_report} defines the interval to report MCMC progress and tropolis acceptance rate (if \code{ga_sampler = metropolis}).}

\item{ga_sampler}{a quoted keyword that specifies the MCMC sampler for the skewness parameter \eqn{\gamma}.
Supported key words are \code{metropolis} and \code{slice}. The default is \code{slice}.}

\item{verbose}{logical; if true, model specification, MCMC progress, and Metropolis acceptance rate (if \code{ga_sampler = metropolis})
are printed to the screen.}
}
\value{
An object of class \code{bgal}. The return object is a list that comprises:

\tabular{ll}{

\code{post_sams} \tab a list of posterior samples. Each element of the list corresponds to an unknown parameter. \cr
\tab \cr
\code{response} \tab the response vector. \cr
\tab \cr
\code{covariates} \tab the covariate matrix. \cr
\tab \cr
\code{prob} \tab the quantile probability. \cr
\tab \cr
\code{beta_prior} \tab the prior for regression parameter \eqn{\beta}. \cr
\tab \cr
\code{starting} \tab a list of starting values. \cr
\tab \cr
\code{tuning} \tab a list of tuning parameters. \cr
\tab \cr
\code{mcmc_settings}: \tab a list of MCMC simulation parameters:
       \code{n_iter}: number of iterations; 
       \code{n_burn}: number of burn-in samples;
       \code{n_thin}: thinning degree. If   \code{verbose = TRUE},
       \code{n_report} defines the interval to report MCMC progress and Metropolis acceptance rate (if \code{ga_sampler = metropolis}).
 \cr
\tab \cr
\code{mh_data}: \tab Null if \code{ga_sampler = slice}; 
                     a vector of Metropolis sampler sampler acceptance and rejection indicators if \code{ga_sampler = metropolis}. \cr
\tab \cr
\code{runtime}: \tab running time for MCMC simulation calculated using \code{system.time}. \cr

}
}
\description{
Function for fitting a Bayesian linear quantile regression with generalized asymmetric Laplace (GAL) errors
       via Markov chain Monte Carlo (MCMC).
}
\references{
Yan, Y., Zheng, X., and Kottas, A. (2024). 
"A new family of error distributions for Bayesian quantile regression".
\href{https://tr.soe.ucsc.edu/research/technical-reports/UCSC-SOE-24-01}{https://tr.soe.ucsc.edu/research/technical-reports/UCSC-SOE-24-01}.
}
\author{
Xiaotian Zheng \email{xiaotianzheng42@gmail.com}
}
