% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bqrgal-package.R
\docType{package}
\name{bqrgal}
\alias{bqrgal}
\title{bqrgal: a package for fitting Bayesian quantile regression based on 
       generalized asymmetric Laplace distributions}
\description{
Bayesian inference for linear quantile regression with a likelihood based on the generalized asymmetric Laplace distribution
}
