% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gal-funcs.R
\name{dgal}
\alias{dgal}
\title{The generalized asymmetric distribution}
\usage{
dgal(y, p0, mu, sigma, ga, logd = FALSE)
}
\arguments{
\item{y}{vector of quantiles.}

\item{p0}{numeric value between 0 and 1.}

\item{mu}{real-valued location parameter.}

\item{sigma}{positive-valued scale parameter.}

\item{ga}{numeric value bounded between L and U, where L is the negative root of 
g(ga) = 1-p0 and U is the positive root of g(ga) = p0.}

\item{logd}{logical; if true, then densities d are given as log(d).}
}
\description{
Density for the four-parameter generalized asymmetric distribution with quantile probability \code{p0}, 
             location parameter \code{mu}, scale parameter \code{sigma}, and skewness parameter \code{ga}.
}
