% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gal-funcs.R
\name{find_ga_ub}
\alias{find_ga_ub}
\title{Find the positive root of \eqn{g(\gamma) = p_0}}
\usage{
find_ga_ub(p0, interval, ...)
}
\arguments{
\item{p0}{numeric value between 0 and 1.}

\item{interval}{vector containing the end-points of the interval to be searched for the root.}

\item{...}{additional arguments to be passed to uniroot}
}
\description{
This function finds the positive root of \eqn{g(\gamma) = p_0} using R's uniroot function
}
