% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gexpit}
\alias{gexpit}
\title{Generalized inverse logit transformation}
\usage{
gexpit(x, a, b)
}
\arguments{
\item{x}{a real number}

\item{a}{lower endpoint of an interval}

\item{b}{upper endpoint of an interval}
}
\value{
The return object is a value between a and b
}
\description{
This function applies a generalized inverse logit transformation to interval [a,b]
}
