% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{glogit}
\alias{glogit}
\title{Generalized logit transformation}
\usage{
glogit(x, a, b)
}
\arguments{
\item{x}{a value between a and b}

\item{a}{lower endpoint of an interval}

\item{b}{upper endpoint of an interval}
}
\value{
The return object is a value between 0 and 1.
}
\description{
This function applies a generalized logit transformation on interval [a,b]
}
