% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bqrgal-generics.R
\name{predict.bgal}
\alias{predict.bgal}
\title{Prediction methods for GAL-model fits}
\usage{
\method{predict}{bgal}(
  object,
  covars,
  probs = c(0.025, 0.5, 0.975),
  err_dens = FALSE,
  predict_sam = FALSE,
  ...
)
}
\arguments{
\item{object}{a fitted object of class inheriting from "bgal".}

\item{covars}{an \eqn{m\times p} matrix of \eqn{p} covariates with which to predict. 
If \code{err_dens} is true, this argument is ignored.}

\item{probs}{a vector of probabilities with values in \eqn{[0,1]} to generate the corresponding quantiles.}

\item{err_dens}{logical; if true, the \code{covars} argument is ignored, and samples or quantiles of the 
posterior predictive error density are returned.}

\item{predict_sam}{logical; if true, posterior predictive samples are returned, 
otherwise only quantiles (specified according to the probs argument) 
of the predictive distribution are returned.}

\item{...}{additional arguments. No additional arguments are supported currently.}
}
\description{
This function obtains out-of-sample predictions from a fitted GAL-based quantile regression model object
}
